requirements.txt
spotipy
pyyaml 
colorama